requirements.txt
streamlit
python-docx
pyyaml
wordfreq

